improve: move main views into respective folders
